Do not cache the template.
Use better SQL query and avoid fetching all products and then filtering them.

Module ps_viewedproducts is currently the heaviest module to use on a product page. On a site with 120 000+ products, it takes more than one second to just load this one module (on average 1100 ms on a quite good server that should be enough for a site like this), thus making the module unusable (no customer wants to wait for two seconds until the page is fully loaded). 

Disabling the caching helps the module to load two times faster. Even with a lots of products to display, creating the template is way faster than loading it's cache, so it's better to not use the cache at all.

About the SQL query - now, the module fetches ALL products from database that are active, then loops through them and compares them with viewed product's ids. You can imagine that this is quite a heavy operation for the database (fetch 120k products and then select like 10 of them?) and for the PHP (so much memory used!). 

Re-creating the query to filter products on the database side, not server side, makes the module exponentially fast. Some benchmark (with cache disabled):
Local testing:
Default template: average 300ms -> 16ms
Angar template: average 650ms -> 100ms

Production site testing:
Angar template: 700ms -> 100ms

**Total speedup (cache + sql):
1000-1500ms -> 50-150ms**
I tried changing the number of products displayed, it's still a lot faster with these modifications.

<!-----------------------------------------------------------------------------
Thank you for contributing to the PrestaShop project! 

Please take the time to edit the "Answers" rows below with the necessary information.

Check out our contribution guidelines to find out how to complete it:
https://devdocs.prestashop.com/1.7/contribute/contribution-guidelines/#pull-requests
------------------------------------------------------------------------------>

| Questions     | Answers
| ------------- | -------------------------------------------------------
| Description?  | Increase module speed up to ten times, especially on large sites
| Type?         |  improvement 
| BC breaks?    |  no
| Deprecations? |  no
| Fixed ticket? | -
| How to test?  | Create site, add products. Download current module, enable profiling. See the time it takes ps_viewedproducts to execute it's hooks. Now apply the changes and compare.

<!-- Click the form's "Preview" button to make sure the table is functional in GitHub. Thank you! -->